refactor(web): extract SelectionPopup from ChatPage into its own module
The floating "Reply ↵" selection action lived inline in ChatPage. Move it to
web/src/pages/SelectionPopup.tsx unchanged so it can grow (a follow-up adds an
"Ask sub-agent" action) and be unit-tested in isolation. ChatPage now imports
it and drops the now-unused CornerUpLeftIcon import.
